+++ resources/META-INF/plugin.xml
@@ -1,7 +1,7 @@
 <idea-plugin>
   <id>com.zt.json-editor</id>
   <name>JSON Editor</name>
-  <version>1.8</version>
+  <version>1.9</version>
   <vendor email="hj_zhengt@163.com" url="https://me.csdn.net/">zhengtao</vendor>
 
   <description><![CDATA[
