+++ src/com/jsoneditor/After.java
@@ -1,0 +1,13 @@
+package com.jsoneditor;
+
+/**
+ * @Description:
+ * @Author: 19043204
+ * @CreateDate: 2020/8/17 15:33
+ */
+@FunctionalInterface
+public interface After {
+
+    void process();
+
+}
